+++ src/components/Popover/Popover.types.ts
@@ -22,6 +22,7 @@ export type Props = Pick<
 	| "disableContentHover"
 	| "disableCloseOnOutsideClick"
 	| "containerRef"
+	| "initialFocusRef"
 > & {
 	children?: React.ReactNode;
 	padding?: number;