+++ CHANGELOG.md
@@ -41,3 +41,4 @@
 - Flyout: better fallback calculations and correctly calculating the gap when positioning
 - Flyout, Popover, DropdownMenu: deprecated forcePosition, added fallbackPositions array and false
 - Flyout: fallback history when no position fits
+- Flyout, Popover: initialFocusRef